Fixed register reading
There was poor data being printed, but the issue was not something that could be solve from smoothing. The issue was that the registers on the BMM150 were not all being read at the same time. The Docs clearly state to read all of the registers at once. Now that the registers are all read at once, the data is of high quality.